SLG2: removed few debug prints and fixed the reported compilation time for texture maps

+++ samples/smallluxgpu2/pathocl/compiledscene.cpp
@@ -675,7 +675,7 @@ void CompiledScene::CompileTextureMaps() {
 	// Translate mesh texture maps
 	//--------------------------------------------------------------------------
 
-	const float tStart = WallClockTime();
+	const double tStart = WallClockTime();
 
 	std::vector<TextureMap *> tms;
 	scene->texMapCache->GetTexMaps(tms);
